+++ app/src/main/java/org/dhis2/uicomponents/map/layer/MapLayerManager.kt
@@ -1,11 +1,11 @@
 package org.dhis2.uicomponents.map.layer
 
 import com.mapbox.mapboxsdk.maps.MapboxMap
-import org.dhis2.uicomponents.map.MapStyle
 import org.dhis2.uicomponents.map.layer.types.EnrollmentMapLayer
 import org.dhis2.uicomponents.map.layer.types.HeatmapMapLayer
 import org.dhis2.uicomponents.map.layer.types.SatelliteMapLayer
 import org.dhis2.uicomponents.map.layer.types.TeiMapLayer
+import org.dhis2.uicomponents.map.model.MapStyle
 import org.hisp.dhis.android.core.common.FeatureType
 
 class MapLayerManager {
