+++ app/src/main/java/org/dhis2/uicomponents/map/managers/TeiMapManager.kt
@@ -3,12 +3,12 @@ package org.dhis2.uicomponents.map.managers
 import com.mapbox.geojson.BoundingBox
 import com.mapbox.geojson.FeatureCollection
 import com.mapbox.mapboxsdk.style.sources.GeoJsonSource
-import java.util.HashMap
-import org.dhis2.uicomponents.map.MapStyle
 import org.dhis2.uicomponents.map.TeiMarkers
 import org.dhis2.uicomponents.map.layer.LayerType
 import org.dhis2.uicomponents.map.layer.MapLayerManager
+import org.dhis2.uicomponents.map.model.MapStyle
 import org.hisp.dhis.android.core.common.FeatureType
+import java.util.HashMap
 
 class TeiMapManager(
     private val mapStyle: MapStyle
